saved 1 doc read by moving emailing guests into its own middleware and piped it into the create-event route
