+++ functions/src/app.ts
@@ -1,6 +1,7 @@
 import * as express from 'express';
 import * as cors from 'cors';
 import { json as bodyParserJSON } from 'body-parser';
+import * as nodemailer from 'nodemailer';
 import { firestore, auth, functions } from './firebase';
 import registerRouter from './routes/register';
 import createEventRouter from './routes/createEvent';
@@ -13,6 +14,7 @@ const app = express();
 app.locals.auth = auth;
 app.locals.firestore = firestore;
 app.locals.functions = functions;
+app.locals.nodemailer = nodemailer;
 
 app.use(cors());
 app.use(bodyParserJSON());
