+++ functions/src/routes/createEvent.ts
@@ -1,6 +1,7 @@
 import { Router } from 'express';
 import createEventGuestsCol from '../middleware/createEventGuestsCol';
 import createEventRecord from '../middleware/createEventRecord';
+import emailEventGuests from '../middleware/emailEventGuests';
 import updateEventGuestsRecord from '../middleware/updateEventGuestsRecord';
 
 const router = Router();
@@ -10,6 +11,7 @@ router.post(
     createEventRecord,
     createEventGuestsCol,
     updateEventGuestsRecord,
+    emailEventGuests,
     (req, res) => {
         res.status(200).json(req.body);
     }
